+++ apps/backend/src/app/api/v1/teams/crud.tsx
@@ -18,6 +18,9 @@ export function teamPrismaToCrud(prisma: Prisma.TeamGetPayload<{}>) {
     display_name: prisma.displayName,
     profile_image_url: prisma.profileImageUrl,
     created_at_millis: prisma.createdAt.getTime(),
+    client_metadata: prisma.clientMetadata,
+    client_read_only_metadata: prisma.clientReadOnlyMetadata,
+    server_metadata: prisma.serverMetadata,
   };
 }
 
@@ -43,6 +46,10 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
         data: {
           displayName: data.display_name,
           projectId: auth.project.id,
+          profileImageUrl: data.profile_image_url,
+          clientMetadata: data.client_metadata === null ? Prisma.JsonNull : data.client_metadata,
+          clientReadOnlyMetadata: data.client_read_only_metadata === null ? Prisma.JsonNull : data.client_read_only_metadata,
+          serverMetadata: data.server_metadata === null ? Prisma.JsonNull : data.server_metadata,
         },
       });
 
@@ -123,6 +130,9 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
         data: {
           displayName: data.display_name,
           profileImageUrl: data.profile_image_url,
+          clientMetadata: data.client_metadata === null ? Prisma.JsonNull : data.client_metadata,
+          clientReadOnlyMetadata: data.client_read_only_metadata === null ? Prisma.JsonNull : data.client_read_only_metadata,
+          serverMetadata: data.server_metadata === null ? Prisma.JsonNull : data.server_metadata,
         },
       });
     });