+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/init-stack
 
+## 2.5.23
+
+## 2.5.22
+
+## 2.5.21
+
+## 2.5.20
+
 ## 2.5.19
 
 ### Patch Changes