+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @stackframe/stack-ui
 
+## 2.5.23
+
+### Patch Changes
+
+- Various bugfixes and performance improvements
+  - @stackframe/stack-shared@2.5.23
+
+## 2.5.22
+
+### Patch Changes
+
+- Team metadata
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.22
+
+## 2.5.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.21
+
+## 2.5.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.20
+
 ## 2.5.19
 
 ### Patch Changes